Refactor application file structure
I did not like that the sub commands that I create would be all
contained into one file. This helps separate code when using flags and
reused components like authentication.  I like this new structure as it
helps separate out the logic in the sub commands and helps to show what
parts are being reused.
