+++ cmd/root/root.go
@@ -1,18 +1,16 @@
-/*
-Copyright © 2024 NAME HERE <EMAIL ADDRESS>
-*/
-package cmd
+package root
 
 import (
 	"os"
 
+	"strings"
+
 	"github.com/spf13/cobra"
 	"github.com/spf13/pflag"
-	"strings"
 )
 
 // rootCmd represents the base command when called without any subcommands
-var rootCmd = &cobra.Command{
+var RootCmd = &cobra.Command{
 	Use:   "spotgo",
 	Short: "A CLI tool for managing your spotify playback",
 	Long: `A CLI tool for managing your spotify playback. This allows you to control your spotify playback from the terminal. 
@@ -26,7 +24,7 @@ This tool is built using the Spotify Web API. You will need to have a Spotify ac
 // Execute adds all child commands to the root command and sets flags appropriately.
 // This is called by main.main(). It only needs to happen once to the rootCmd.
 func Execute() {
-	err := rootCmd.Execute()
+	err := RootCmd.Execute()
 	if err != nil {
 		os.Exit(1)
 	}
@@ -48,7 +46,7 @@ func init() {
 
 	// rootCmd.PersistentFlags().StringVar(&cfgFile, "config", "", "config file (default is $HOME/.gosp.yaml)")
 
-	rootCmd.SetGlobalNormalizationFunc(wordSepNormalizeFunc)
+	RootCmd.SetGlobalNormalizationFunc(wordSepNormalizeFunc)
 
 	// Cobra also supports local flags, which will only run
 	// when this action is called directly.
